.github/workflows: Don't execute FTPUpload job on workflow cancel
The FTPUpload job is responsible for storing the job artefacts of all
jobs regardless if the jobs passed or failed.

But we don't need to store the artefacts if the job was cancelled.

This avoids a failed job in case of cancellation and also the nagging
email from github actions.