SF_PreprocessInput: Remove trailing whitespace to avoid bugging out with variable assignments
The newly added test case fails without the fix in SF_PreprocessInput.

We search for all variable statements in the preprocessed
code in SF_GetVariableAssignments called via SF_ExecuteVariableAssignments. And we abort the variable search, as there is no explicit boundary between variables and operations, once we could not read a line as variables.

And trailing whitespace in a line is not a valid variable statements,
therefore we bug out with a message like "Only one operation are allowed"
or other assertion messages from SF_FormulaExecutor.

We now remove trailing whitespace in the SF code preprocessor.