SF_GetDefaultFormula: Don't use cursors
The initial SweepFormula design had a prominent place for the cursors
operation. The idea was to allow users to extract data from arbitrary
ranges. But as soon as we had epoch support back in 9cfbd579 (SweepFormula
add epochs command, 2021-09-03) we realized that one wants to usually just
extract an epoch and not an arbitrary range.

The special cases which is exploited in the cursors(A, B) code is that
when there are no cursors present it returns the whole range.

But we can as easily return the whole range via [0, inf].