SF: Exchange order of querying wave data from JSON wave notes
With the fix in the previous commit we can now be sure that when we
request numeric data we always get it.

It therefore makes sense to ask first for numeric and then for text. As
that is also correct for cases of "inf"/"-inf"/"NaN" which is numeric but
is also parsed correctly as text.