Propagate changed stimsets better
Since the introduction of combine epochs for the wavebuilder in 9b809878
(WaveBuilder: Add new epoch type Combine, 2015-11-12) we need to maintain
a wave with the shorthand<->full name mapping. This mapping is updated in
WB_UpdateEpochCombineList.

Now that function was not called when loading stimsets from NWB and not
after starting the wavebuilder panel which resulted in errors trying to
recreate combine epochs.

This is now fixed. In order to make the fix less hacky we also introduced
WB_UpdateChangedStimsets which also calls
DAP_UpdateDaEphysStimulusSetPopups which needs to be called in the same
cases.