Fix: Testpulse calculation failed if TP data is NaN
In the case the test pulse data is completely NaN WaveStats can not
determine V_minRowLoc and V_maxRowLoc and returns a -1 in both.
This index was used unchecked and resulted in a RTE for invalid
wave index access.

This is now checked properly and NaN is returned for "instantaneous".

since
c5ed1686 (TP_TSAnalysis: Change the way we calculate the instantaneous resistance, 2022-09-02)