MIES_AnalysisFunctions_PatchSeq.ipf: Remove fallthrough to PRE_SET_EVENT
In 76b7cd2f (Patch Seq: Make it work with indexing, 2018-06-11) we
introduced support for using indexing with the patch seq analysis
functions. We did that by executing the code we can in PRE_SET_EVENT and
everything else in PRE_DAQ_EVENT. The added fallthrough means that we
execute the code in PRE_SET_EVENT also in PRE_DAQ_EVENT.

It is not clear why we did it back then like that, so let's remove the
fallthrough.

We can then also remove the DAScale setting for PSQ_DAScale in the
PRE_DAQ_EVENT.

Close #1192