Tests/CheckDashboard: Ensure we don't have legacy early abort DAC messages present
Via CommonAnalysisFunctionChecks which we call at the end of each analysis
function test we already enforce that we have a valid dashboard pass/fail
message.

But since 8657a249 (AD_GetPerSweepFailMessage/AD_GetBaselineFailMsg:
Handle user aborted sweeps better, 2023-05-30) we try harder to also
detect early stopped sweeps prior to 87f9cbfa (DAQ: Add stopping reason to
the labnotebook, 2021-05-13).

But we don't want to have the legacy sweep stop message present in any of
the tests of our analysis functions. As that always indicates a gap in the
dashboard evaluation.