Tests/CheckRangeOfUserLabnotebookKeys: Prefer sweepNoRef
We want to get the list of the RAC of the passesd sweepNoRef not of
sweepNo which is zero at this point.

This is not an issue in practice as we only ever acquire one RAC in the
users of the function.

Bug introduced in 4f660b6a (Tests for analysis functions: Add generic
range checks for labnotebook entries, 2022-03-22).