SWS: Add SWS_ProcessDATTLChannelsOnEarlyAcqStop function to retrieve acq time
For saving we also have to write the correct times into the epochs wave,
thus the actual acquired time and planned time must be determined.
Both are typically the same except when the acquisition was aborted early.

The function determines both times.

Additionally with the prefill of the DA/TTL channels in the scaledDataWave,
the samples that the acquisition did not reach in case of early abort
are filled with NaN.

Based on the acquisition time also a sanity check is done, where the
ADC channel is checked that the samples in the interval where no
acquisition took place are still NaN (the ADC channels are prefilled with NaN).

EP: Use times as arguments for EP_WriteEpochInfoIntoSweepSettings

The times are determined in SWS_SweepSaving previously by
SWS_ProcessDATTLChannelsOnEarlyAcqStop to wrap some related
functionality in that utility function.