LOG_AddEntry: Readd timestamp
The template function LOG_GenerateEntryTemplate had its ts object with the
timestamp removed in 2fab915a (LOG_GenerateEntryTemplate: Remove ts
object, 2023-05-24). The idea was the don't want to have the timestamp in
the template for the log writters in the ITCXOP and the ZeroMQ XOP.

But we forgot that we also use that function via LOG_AddEntry and this
needs have a timestamp.

So let's add it back here.