PSQ_GetLastPassingDAScale: Don't overwrite key inside loop
In eee70af7 (PSQ_GetLastPassingDAScaleSub: Check the value of set QC,
2021-06-22) we have moved the check for the operation mode inside the foor
loop. But that change also overwrote key inside the loop. The result is
that if the last passing sweep has a different operation mode than
expected we don't pass the setQC check again as key is now wrong on the
next iteration.

Fix that by introducing a separate string for that.

The bug was asymptomatic as every sweep has an operation mode, and IVSCC,
the main user of these analysis functions, always runs the analysis
functions in the same order.