EP: Fix that end of epochs are limited to end at last data point
last data point is stopCollectionPoint - 1, where the Stimset ends,
even if it is from definition stretching to the range between the
last and the last + 1 data point.
The range is cut off at the last data point because otherwise the data
would include "undefined" time (e.g. no stimset defined for the
last + 1 data point)