EP: When adding TP epochs do not redo parts of the calculation
With low DA frequencies like 10 kHz the re-calculation of the TP boundaries
for the TP epochs from the initial times and fraction resulted in
boundaries that were off by one sample point when taking the actual
inserted TP as reference. This resulted that the test for the TP amplitude
in the DA data based on epoch boundaries failed.

Fix:
- use the already calculated values for the TP epochs
- these TP epochs are derived from the actual filling of the DA channel data

Testing of that fix is covered in EP_EpochTest16 and for Sutter hardware
in every epoch test that has inserted TP enabled.