Tests: Adaptation to Epoch tests for strong variation in DA frequency
Stimset epochs are defined through arbitrary points in time. The actual
data is then created in the wavebuilder and decimated to the DA frequency
in the DAQDataWave. The tests for epochs are including an amplitude check
that tests for the easy case of pulse waves if the defined epochs and the
amplitudes in the DA wave fit. For that check the boundaries of the epochs
in sample points of the DA wave need to be determined.
For that a rather simple approximation was used that worked well for
typical sample frequencies of ITC and NI hardware for one or two channels
(about 125 - 250 kHz).
Tests with sutter hardware that reaches maximum DA frequencies to 10 kHz
failed because of the approximation for the sample point determination
was not suitable for such low frequencies any more.

This change fixes that by using a better approximation for the test.
It has to be taken into account that first the stimset epochs are put as
data into the wave created by the wavebuilder at 200 kHz. Then the data
is decimated when transferred to the DAQDataWave to another sample
frequency. As a result the actual epoch boundary in the samples of the
DA channel wave is shifted regarding to the definition in the epoch.

The approximation works like that:
For the transfer of the arbitrary epoch start or end time a grace time of
1.5 times the sample interval of the wavebuilder is added/subtracted
(7.5 ms for 200 kHz).
With that time the next/previous sample point in the DA channel wave is determined.
If the time is very close to a sample point then the next/previous sample
point is takes because then the sample point is where the signal amplitude
changes from one to another epoch.
If the time is between two sample points then it is rounded up/down to the
next sample point.

Between the resulting sample positions the amplitude is checked.

The fix purely adapts TestEpochsMonotony where the amplitude check is done.

This is verified in EP_EpochTest16 where at 10 kHz is acquired, whereas
the remaining Epoch tests (when using ITC, NI) use higher frequencies.
For sutter hardware it is verified by every Epoch test that includes
a pulse train stimset as sutter hardware runs a maximum DA frequency of 10 kHz.