WB: Add function to determine indices from startTime and duration
- use it for pulse lengths in pulse trains

Previously:
startIndex = floor(start / interval)
endIndex = floor((start + duration) / interval)

this has two problems:

1. startIndex does not choose the closest sample point to the time start.
e.g. if start = 1.9 and interval = 1 then startIndex would be 1 with
an error of 0.9 whereas 2 is much closer with an error of 0.1 only.

2. The amplitude should be kept for at least the given duration.
Thus, if the error in time introduced by floor for endIndex is (in absolute)
larger than for startIndex the resulting duration in the output is actually
smaller than the given duration in time.

The changed approach uses for startIndex round() to determine the closest
sample point. The time referenc epoint for start is then then taken from
the just chosen sample point and ceil() is used to have a resulting
duration that is always at least as long as the duration given as argument.