Tests/PrepareForPublishTest: Update XOP logging template
While in MIES production usage, the logging template is set from the
various hook functions in MIES_IgorHooks.ipf, these are partially not run
for automated testing as the MIES procedures reside there in "User Procedures" and
not "Igor Procedures".

So let's setup the logging template in preparation of the publish
test.