BSP_UpdateHelpNotebook: Call on-demand
This fixes an issue where the SweepFormula help file (ihf) was not
present on loading saved experiments which had a databrowser/sweepbrowser
open.

The reason was that we only loaded the ihf file on browser creation and
that does not happen when loading an experiment.

Now BSP_GetSFHELP loads the ihf file on-demand.