AD_FillWaves: Fix calculation of failed sweeps for Rheobase
PSQ_Rheobase does not have a per Sweep QC labnotebok setting. Therefore we
decided in 23f0714f (Analysis functions dashboard, 2018-05-06) to use, for
passing set QC, the last spiking/non-spiking sweeps as passed and
everything else as failed.

Unfortunately it was not anticipated that only two sweeps are present and
in that case we have one sweep both failing and passing as Duplicate/RMD
clips its passed indices.

By using the set difference we avoid this issue.

Close #2017