AB: bugfix LoadDataWrapper excluded loaded waves if the casing was different
LoadDataWrapper does a check if it can find the loaded waves after the load.
This check was case-sensitive. Thus, it failed if the input wave name was
e.g. lower case and the actual loaded wave was upper case named.

Fix: Make the check case-insensitive to be in line with general object name
handling in Igor Pro

since d13a6b7