WB: fix recreation of dependent Stimsets from Stimset parameter waves
The WB has an issue, where non-existing stimsets are not treated as
needing an update. Thus, combine type stimsets that depend on another
stimset being created first can not be created because the other stimset
is not being created from the parameter waves.

since 792d0d24 (WB: update stimset on changes, 2016-10-31)