DC: Add fallbacks in recreation for MIES version < 0.7
The very early MIES versions did not store the offset of the stimset in
the DA wave. The offset is required for epoch recreation because it allows
to split the DA wave into the onset, stimset and trail region.

1.
Added a fallback algorithm that exploits the fact that the trail is
a zero line. So the position of the last feature of the DA wave is
matched against the last feature of the decimated stimset.

Then the decimated stimset is compared to the DA wave content in a
-2 to +2 range from the initial guess.

If that fails then the whole range from 0 to the initial guess is searched
for an optimal match between decimated stimset and DA wave content.
The intense search is time consuming and a "last option" fallback.
Checking available test data it was never required, so it should be
a very rare case its necessary (or may hint to a flaw in the method for
the initial guess).

2.
For the special case that the stimset itself is a zero line, the DA wave
is checked if there is a feature, that can only be a TP.
From that TP and its baseline symmetry the offset of the stimset start is
determined.