UTF_AsynFrameworkTest.ipf: Correct out-of-order check
Since the introduction of the tests in 14cd1ff8d6 (Add Async Framework,
2018-09-17) we tried to verify that the execution really happened out of
order. The fix in b8ae7c65ce (Asyn: Fix WARN(1) in tests for orderless
execution, 2022-11-03) did not fix it though as we actually need to check
that the returnOrder array is not ascending.

As the workload in RunGenericWorker already does a randomized workload we
feel lucky that it is never sorted and check that.