Tests: Add test for checking user epochs in historic data
- renamed TestEpochRecreation to TestEpochRecreationFromLoadedPXP to
  prevent naming confusion with the utility function of the same name in
  UTF_HelperFunctions.ipf
- Limited TestEpochRecreationFromLoadedPXP to test only for pxp files as
  the testing approach does not work with nwb files.
- Added test TestEpochRecreationShortNames and related utility function
  ExportEpochsFromFileToDF to create additional reference data.
- TestEpochRecreationShortNames tests for existence of epochs in
  historic data.
  1. all epochs present in the historic data reduced by the in epoch recreation
    supported user epochs must be present in the recreated epochs
  2. all user epochs from recreated epochs must exist in the epochs of the
    historic data

Some historic data does not have existing user epochs. For these files
recreated epochs can be exported with the utility function ExportEpochsFromFileToDF
to a data folder in HistoricData.pxp.
For these files the refEpFolder has to be set in TestEpochRecreationShortNames.
Then the second part of the TestEpochRecreationShortNames tests effectively,
if the current implementation of epoch recreation recreates all user epochs
that were recreated at the time the export with ExportEpochsFromFileToDF
was done, i.e. changes in the current code that remove previous user epochs
are found.