Util: Adapt FindRange to fix LNB sweep ordering issue
With the introduction of the PRE_DAQ_EVENT in analysis functions the analysis
function could write LNB entries with the speculative next sweep number, even
if the sweep was then not acquired, e.g. due to failing DAP_CheckSettings().
As a result, following LNB entries from TP were added with the last sweep number.
The LNB order looked then like this.

sweepNo
N
N
N + 1 <- from analysis function with speculative sweep number
N *
N *
N + 1 <- Actual N + 1 sweep that was acquired
N + 1
...

Entries in the LNB are found by looking for a consecutive block starting from the back.
So the found block for sweep N would result in the range marked by *. The preceding entries
were ignored as they were unreachable.

This commit fixes this behavior for reading the LNB.
The fix is the following:
The determination of the block range ignores entries with entry source type UNKNOWN.
Entries from analysis function have that entry source type.

When searching for the specific entry in that range the search ignores rows that do
not have the correct sweep number.

- increased version of LBRowCache wave
- added test that checks specifically against the case described above from an user experiment
  Basic.pxp was updated with a LNB part for this test.

since
76b7cd2f (Patch Seq: Make it work with indexing, 2018-06-11)