BSP_RenameAndSetTitle: Keep automation suffix
In 7ef8625c8 (Databrowser/SweepBrowser: Add mode property, 2022-10-12) we
introduced the user and automation browser types. For the automation
browser we changed the title via BSP_SetBrowserType.

But 6bdae2689 (BSP_RenameAndSetTitle: Make the title more informative,
2022-11-06) broke that as it was rewriting the title again after browser
locking to a device.

So we now set the title only in BSP_RenameAndSetTitle.