PSQ_DS_GatherFrequencyCurrentData: Distinguish PRE_SET_EVENT and POST_SWEEP_EVENT usage
In PRE_SET_EVENT we fit the rheobase and supra data. For that we take the
starting point of the E1 epoch of the data but use the adaptive supra E1
epoch length.

For POST_SWEEP_EVENT we just want to take the E1 epoch length.

This fixes that the assertion adaptiveEpochLength <= supraEpochLength was
triggered for prematurely stopped adaptive supra sweeps when evaluated in
POST_SWEEP_EVENT.