Tests: Add special adaptation for U_RA_UD epochs
The adaptation removes the check for the end time of the U_RA_UD epochs.
For that the end time is set to the end time of the reference epoch wave.

The reason is that the approximation in the determination of the end time
depends on the AD data acquired after the black out time. The approximation
treats the first appearance of a AD value that is non-zero as end time.
However, if the actual AD values acquired after the black out are zero,
then the approximation includes these in the black out range.
This the end time from the approximation does not have to be exactly
equal to the original zero filled range in the AD channel wave.
Thus, it has to be removed from the equality check.