SFH: Added a variant of SFH_ResolveDatasetElementFromJSON that also returns the datasets datatype
The datatype is in the note of the wave reference wave and not in the data wave itself.
Thus, the returned wave of SFH_ResolveDatasetElementFromJSON does not contain this information.

SFH_ResolveDatasetElementFromJSONAndType was added that returns both. This allows operations
with loose argument order if all the arguments are typed.