SF: Add support for operation results in arrays
Operation results can be any wave type. Previously arrays only supported
numerical and text elements.

For operation results the returned result is now handled as follows:
- if the result is a primitive numeric result it is directly used as array element
- if the result is a primitive textual result it is directly used as array element
- if the result is not of type numeric or text it is wrapped as a wave reference as
  string (@WREF prefix) and is, thus converted to a primitive textual result
- if there are multiple results it is wrapped as wave reference as well

The wrapping as wave reference requires that the current subarray is of type text.

The previous quick path for numerical and textual waves was removed in favor of
an element based evaluation for subarray. Each element is resolved, also operations and variables.
The type of the output (out/outT) is determined on first creation and verified for type collisions
for each further element.
The array dimensionality increase is handled depending on the resolved data. For scalars the array
is not expanded.