PSQ_GetSweepFormulaResultWave: Don't check the selected sweeps
Ever since the creation of the function in fab0fcf73
(PSQ_GetSweepFormulaResult: Extract out helper function, 2022-02-22) we
check that the requested sweep is one of the displayed sweeps.

But that does not make sense as the user might change currently displayed
sweeps. So instead we now don't check the sweep number, this requires that
the last store is the one which we want to query. This is the case for all
current function calls of PSQ_ExecuteSweepFormula/PSQ_GetSweepFormulaResult/PSQ_GetSweepFormulaResultWave.

Close #2129