AB_LoadSweepsFromExperiment: Handle no config waves gracefully
It can be the case that we have sweeps in the labnotebook but no sweep
config waves and data. Although this is an edge case, we want to handle
that gracefully. The current implementation results in an RTE as
SweepConfigDFR was invalid after AB_LoadSweepConfigData returned with no
config waves loaded.

Fix that by returning early and add a test for that.