SB: Add testing assertion in SB_GetSweepIndexFromMap for multi device usage
In principle the sweep browser supports multiple devices/experiments.
However not all sweep browser functions imlement a suitable API for that.
SB_GetSweepIndexFromMap is such function. When running tests it asserts now
if sweep numbers appear more than once, which can happen when multiple
experiments are loaded and this function is called.