LNB: Change LNB column determination to use a sorted keyWave in the same DF
FindDimlabel for searching to correct Logbook column is slow. A better approach is to use a
sorted wave with the keys and do a binary text search. This change replaces FindDimlabel with
such an approach.

A wave getter is added that retrieves two temporary waves, one with the sorted keys and another
with the column indices of these keys (text and numeric wave).
The waves are located in the :Temp subdatafolder of the logbooks wave folder.

For better debugging the sortedKeys and the indices are stored as permanent waves.

In the tests PrepareLBNNumericalValues and PrepareLBNTextualValues function was added
because for already existing logbook wave templates the creation of the temporary waves
would happen outside of the current data folder. With wave tracking this counts as wave leak.
Instead the working logbooks for a test case are duplicated in the current DF and used from
there for the tests.

Considerations were:
- caching is not used because the cache lookup time is too slow