CleanupOperationQueueResult: Add and use it
When using the operation queue via Execute/P there is the option to
suppress an error popup via /Z. But this suppression creates a global
variable V_flag.

The function CleanupOperationQueueResult allows to remove that global
variable if it exists. And it also has to use the operation queue so that
it is executed later than the original operation queue command.