LNB: fix initialisation of column labels for new LNBs
LNBs are created with INITIAL_KEY_WAVE_COL_COUNT numbers of columns that
is 5 as of this commit. However only labels for 4 columns are set, leaving
the last column without a dimlabel.
Also there is already a list of keys used for the keyWave that can already
be used as DimensionLabels for the columns.

As soon as data is acquired, i.e. the first LBN entry is written through
ED_FindIndizesAndRedimension then all keys are set as DimLabels with
LBN_SetDimensionLabels(key, values, start = (rowIndex == 0 ? 0 : numKeyCols))
where a special case for no row entries is done.
This sets the former empty column DimLabels.

since
5c9a5e4c (Add "Acquisition State" as default labnotebook entry, 2020-11-27)