PSQ_DS_FitFrequencyCurrentData: Avoid RTE with too few data
In the singleFit case we remove all data at the front and only leave
numPointsForLineFit left.

But if we don't have enough elements to start with, we currently RTE out
at the DeletePoints.

Avoid that by only doing DeletePoints if we have enough points. Not having
enough points already finished earlier a few lines below.