SetDaScale/SetDAScaleModOp: Allow to respect DAScale limits
When limitCheck is true we now return 1 when setting the new DAScale value
would be out of range for the current external command sensitivity of the
MCC amplifier.

This also requires that we now pass in the sweep number.

The default of limitCheck is still false. Only when all analysis functions
are ported, can we flip the default.