PSQ_DaScale: Enhance adaptive mode
We now use the previous acquired (RhSuAd) sweeps for judging if we have enough
passing sweeps with f-I slope QC. For this we need to store the list of passing
RhSuAd sweeps together with their f-I slope QC status.

Calculating the f-I slope QC for these sweeps posed the problem that we have to
distinguish sweeps which are left to the maximum slope (usually smaller but not
passing) to the ones to the right. We now take that into account in
PSQ_DS_CalculateReachedFinalSlope.

This made it also required to keep f-I slope values which stem from duplicated
apfreq values. These are now added with the value -inf (PSQ_DS_SKIPPED_FI_SLOPE).

These changes now can create the situation that we start a adaptive
suprathreshold measurement but due to previous RhSuAd sweep already passing
enough for our numSweepsWithSaturation we just repeat the last DAScale value
and are done.

The function PSQ_DS_GetLabnotebookData does all the heavy lifting for getting
the desired labnotebook values for the RhSuAd data and the current SCI.