Wavebuilder/DaEphys: Don't allow stimsets to have non-finite entries
When using delta modes like log/power it is pretty easy to create stimsets
which have NaN/Inf/-Inf values. But this does not make sense as the
hardware can't apply these infinite values and very likely the user also
doesn't want them.

So let's prevent saving stimsets with them and add a check in
DAP_CheckStimset.

The stimsets Ref1_c_DA_0 and Ref1_e_DA_0 from the tests in
UTF_WaveBuilderRegression.ipf had to be fixed as well.