.github/workflows: Prefer on.push
We currently use on.pull_request for running CI on feature branches. This
creates a merge commit against the base branch to run the CI against. This
means that the git revision of that CI run is not the HEAD of the branch
PR but a transient revision.

This is bad for traceability for us, as we sometimes have long-living
branches in semi-production.

By using on.push we can avoid these issues. But we also have to use other
variables as the github.event has different entries for push events.

Close #2117