Fix: RecreateMissingSweepAndConfigWaves handle non-missing waves properly
For the case of non missing sweep or config waves the waves were
retrieved by a simple WAVE/Z in the current DF, but the waves are actually in
deviceDataDFR.

If both waves exist then the validity check assumes that the config wave
is the latest version, but that could be not the case.
Now the actual version of the existing config wave is given, such that the
check verifies the wave geometry.

since 6a432cb7 (Add code to allow recovering deleted sweeps, 2021-04-19)