NWB: Add ChunkedCompression for Stimset save in NWB_PrepareExport
By default the sweep data and stimsets used in that sweeps are saved
with ChunkedCompression. Also when exporting all data then the Stimsets
are saved with ChunkedCompression.

Thus, it makes no sense to not use ChunkedCompression as well when
appending the stimsets when a NWB file is created through the
acquisition from a DAEPhys panel.