ASYNC: If TS is disabled then free DF are now moved to a global DF
The underlying issue is that for some caching key determination we use
WaveModCount that does not support free waves when executed from the
main thread.
This issue comes up e.g. when the LNB is accessed due to the key wave caching.

The solution is to move the prepaed DF to a global location before calling
ASYNC_Worker. The global DF is created as a trash* DF in the MIES DF.
Since certain actions, like TP acquisition would create a massive amount of
these temporary DFs, the trash DFs are cleaned up whenever
a new ASYCNC workload is pushed. This is good enough for debugging
purposes.