SF_PreparePlotter: Remove controls/draw elements on existing panels
In dc9ed87e4 (SF_PreparePlotter: Don't kill and recreate the main panel
for SF_DM_SUBWINDOWS, 2024-06-19) we switched from killing the
sweepformula plot to clearing it.

But we forgot to remove additional controls and draw elements added by
e.g. PSX.

And now we also have to explicitly remove TraceUserData (TUD) as these are
not cleaned up automatically anymore, as killing an embedded subwindow
does not trigger the main window hook.