PS_DS_AD9: Fix it
This test uses PSQ_TEST_VERY_LARGE_FREQUENCY to use a frequency which
results in a very large fit slope. The update of the JSON XOP extended the
range of valid values, so we also need to use a larger value here.

The removed finite check in CheckRangeOfUserLabnotebookKeys is unfortunate
but should not be a problem.