TraceValueDisplayHook: Add it
For plots like the labnotebook or the analysis function parameters it is
easily possible that one has multiple traces with more or less the same
x and y values. And for these cases it is difficult to distinguish the
entries.

We now have a custom window hook which creates a tooltip with the names
and values of the labnotebook/analysis function parameter.