BP: Adapt CreateTiledChannelGraph for multi-device-exp support
CreateTiledChannelGraph is used in the SweepBrowser and the SB
supports dispalying data that is loaded from multiple experiments
and could be acquired from multiple devices.

This information needs to be properly included in the trace user data
as sweepformula needs it to be able to refer to the correct source
sweep data from the trace user data.

The mapIndex in the sweepBrowser sweepMap is unique for a specific
device/experiment combination.