Bugfix: SF use proper handling of sweepMap for SB
problem described in https://github.com/AllenInstitute/MIES/issues/2151

SF requires to retrieve the correct LNBs and sweepDFR for sweeps.
Through the analysisbrowser sweeps from multiple devices and/or
experiments can be loaded. This must be handled correctly or wrong
data is retrieved/analysed.

Thus, SF functions that retrieve sweep data are adapted to handle that
correctly. For selections a forth column was added to store the mapIndex
of the sweepMap that is unique for a sweepNo, experiment, device combination.

For graphs from SB the mapIndex is used to retrieve the correct LNB, SweepDF.

** THIS BUGFIX IS LIMITED TO SWEEP HANDLING IN SWEEPFORMULA **

Other parts of MIES may be still unaware of multiple experiment/devices in the SB.