MIES_SweepFormula_PSX.ipf: Handle sweepbrowser labnotebooks correctly
In the same spirit as d5b560f5f (Bugfix: SF use proper handling of
sweepMap for SB, 2024-10-04) but for PSX.

We also need to include now the experiment name in the combo key to make
it unique. And for the stats equivalence data we also include the
sweepmapindex next to the sweepNo.