PA_SpikePositionsForNonVC: Fix error in execution
In 2bc24dac1 (PA_SpikePositionsForNonVC: Add missing /Q, 2024-08-01) we
added a plain /Q  to FindPeak. But the author did not realize that for
this operation the /Q implies not erroring out (other operations do that
with /Z).

So the function currently always bugs out when it finds a peak.

By dropping the RTE error handling we can use it correctly.