FindRange: Fix querying data from NaN sweep
When we look for data in the NaN sweep, these are the labnotebook entries
before the very first sweep and don't care about the entry source type
(UNKNOWN_MODE) we did not find anything.

The reason is that we are searching from the back for a continous range of
entries with the given source type and sweep number.

But if we have empty entries at the back of the labnotebook, which is very
likely, we always only search in these. But as these are empty rows, we
don't find anything.

For finite sweep numbers this is an optimization at the FindIndizes call
only.

The check for non-finite endRow is needed as we don't upgrade labnotebooks
from the analysis browser.