PSQ_DAScale (Adapt): Round calculated DAscale to integer values
We don't need the sub pA precision. In PSQ_DS_GatherOvershootCorrection we
now also check if the calulated value is the same as the before/after
value, as this can now happen with rounding.

In PSQ_DS_CalculateDAScale we also use ceil for the too large/small cases
so that we always get a different value.