.github/workflows/test-igor-rebase-exec-workflow.yml: Fix it for the release branch
In c3e7dc3a (.github/workflows: Add job to compile test each commit,
2023-12-22) we introduced a compilation test of each commit in the branch.
But we assumed that the base branch, the branch we want to merge into, is
origin/main. This is not true for the release branches.

As there is no way to gather the base branch for on.push workflows [1] we
added some code to get the release branch with the highest version and
then check if we are an ancestor of that or origin/main.

[1]: https://github.com/orgs/community/discussions/26243#discussioncomment-3250924