UpgradeLabNotebook: Fix upgrade of NOTE_INDEX for textualValues waves
We can't search for NaN here but need to search for an empty string. This
was not caught by the test as that used an empty wave and for these
FindValue does not complain.

The fallout from this bug is an empty row range in the labnotebook. Which
should not matter.