Databrowser/SweepBrowser: Fix some more sweep/index confusion
When loading multiple experiments from the analysis browser, we expect
that the total number of sweeps in the display is correct.

This was not the case as we used the last sweep number and not the number
of sweeps.

In addition it was not possible to select sweeps from the popupmenu from
later experiments as it always selected the same sweep from the first
experiment.