GetZeroMQXOPFlags: Turn off logging
With the every TP PUB message we start to get large logfiles. A rate of
45MB/h was reported.

We added the logging of the full in/out messages in 4ba30ba9e
(StartZeroMQSockets: Enable ZeroMQ logging, 2021-04-06) to be able to
debug issues more easily. While that is still a loudable goal, we also
need to ensure not to waste resources on log files.

So let's turn it off for now.