MIES_AmplifierInteraction.ipf: Rework AI_UpdateAmpModel
It is now possible to pass in an amplifier function constant and a clamp
mode instead of a control name. This prepares for a future commit where we
make the setter part of AI_SendToAmp static and force all users to use
AI_UpdateAmpModel instead.

And we also pass in the value directly from the amplifier GUI controls.