Fixes in installer script for NSIS 3.10 update
The plugin ExecCmd is not supported and not working anymore.
The replacement plugin is ExecDos.

While the API of ExecDos looks similar to ExecCmd it
behaves in a lot aspects differently.

Also some documented functionality like "application output to log file"
was not possible to get working.
Environment variables as part of the command line are not supported.
(Need to be resolved before)

Thus, the method used now is where the plugin puts the output on the stack.
As workaround for the piping the stack is written to a file.
Then the second command is called in a separate call.